[REF] sign_oca: Remove legacy code
