+++ sign_oca/__manifest__.py
@@ -34,7 +34,6 @@
         "web.assets_backend": [
             "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.xml",
             "sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.xml",
-            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.xml",
             "sign_oca/static/src/elements/elements.xml",
             "sign_oca/static/src/scss/sign_oca.scss",
             "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.esm.js",
@@ -42,8 +41,6 @@
             "sign_oca/static/src/elements/text.esm.js",
             "sign_oca/static/src/elements/signature.esm.js",
             "sign_oca/static/src/elements/check.esm.js",
-            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.esm.js",
-            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf_action.esm.js",
             "sign_oca/static/src/components/"
             "sign_oca_pdf_common/sign_oca_pdf_common_action.esm.js",
             "sign_oca/static/src/js/*.js",
@@ -52,19 +49,15 @@
         "web.assets_frontend": [
             "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.xml",
             "sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.xml",
-            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.xml",
             "sign_oca/static/src/elements/elements.xml",
             "sign_oca/static/src/scss/sign_oca.scss",
             "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.esm.js",
             "sign_oca/static/src/elements/text.esm.js",
             "sign_oca/static/src/elements/signature.esm.js",
             "sign_oca/static/src/elements/check.esm.js",
-            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.esm.js",
             "sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.esm.js",
             "sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.xml",
             "sign_oca/static/src/scss/portal.scss",
-            "sign_oca/static/src/js/*.js",
-            "sign_oca/static/src/xml/*.xml",
         ],
         "sign_oca.sign_assets": [
             "sign_oca/static/src/scss/sign.scss",
