--- sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates xml:space="preserve">
-    <t t-name="oca_sign_oca.SignatureButtons">
-        <div class="o_sign_oca_buttons" role="toolbar" aria-label="Main actions">
-            <button
-                type="button"
-                class="btn btn-primary o_sign_oca_button_sign"
-                accesskey="s"
-                t-if="to_sign"
-            >
-                Sign
-            </button>
-        </div>
-    </t>
-</templates>
