+++ sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.esm.js
@@ -2,16 +2,19 @@
 
 const {App, mount, useRef} = owl;
 
-import SignOcaPdf from "../sign_oca_pdf/sign_oca_pdf.esm.js";
+import SignOcaPdfCommon from "../sign_oca_pdf_common/sign_oca_pdf_common.esm.js";
 
 import env from "web.public_env";
+import {registry} from "@web/core/registry";
 import {renderToString} from "@web/core/utils/render";
 import session from "web.session";
 import {templates} from "@web/core/assets";
+const SignRegistry = registry.category("sign_oca");
 
-export class SignOcaPdfPortal extends SignOcaPdf {
+export class SignOcaPdfPortal extends SignOcaPdfCommon {
     setup() {
         super.setup(...arguments);
+        this.to_sign = false;
         this.signOcaFooter = useRef("sign_oca_footer");
     }
     async willStart() {
@@ -39,9 +42,19 @@ export class SignOcaPdfPortal extends SignOcaPdf {
             $("#sign_oca_button").prop("disabled", true);
         }
     }
-    postIframeFields() {
-        super.postIframeFields(...arguments);
+    postIframeField(item) {
+        /* eslint-disable */
+        var result = undefined;
+        /* eslint-enable */
+        if (item.role_id === this.info.role_id) {
+            var signatureItem = super.postIframeField(...arguments);
+            signatureItem[0].append(
+                SignRegistry.get(item.field_type).generate(this, item, signatureItem)
+            );
+            result = signatureItem;
+        }
         this.checkFilledAll();
+        return result;
     }
     async _onClickSign() {
         const position = await this.getLocation();
@@ -68,6 +81,77 @@ export class SignOcaPdfPortal extends SignOcaPdf {
                     window.location.reload();
                 }
             });
+    }
+    _trigger_up(ev) {
+        const evType = ev.name;
+        const payload = ev.data;
+        if (evType === "call_service") {
+            let args = payload.args || [];
+            if (payload.service === "ajax" && payload.method === "rpc") {
+                // Ajax service uses an extra 'target' argument for rpc
+                args = args.concat(ev.target);
+            }
+            const service = this.env.services[payload.service];
+            const result = service[payload.method].apply(service, args);
+            payload.callback(result);
+        } else if (evType === "get_session") {
+            if (payload.callback) {
+                payload.callback(this.env.session);
+            }
+        } else if (evType === "load_views") {
+            const params = {
+                model: payload.modelName,
+                context: payload.context,
+                views_descr: payload.views,
+            };
+            this.env.dataManager
+                .load_views(params, payload.options || {})
+                .then(payload.on_success);
+        } else if (evType === "load_filters") {
+            return this.env.dataManager.load_filters(payload).then(payload.on_success);
+        } else {
+            payload.__targetWidget = ev.target;
+            this.trigger(evType.replace(/_/g, "-"), payload);
+        }
+    }
+    async getLocation() {
+        if (!this.info.ask_location || !navigator.geolocation) {
+            return {};
+        }
+        try {
+            return await new Promise((resolve, reject) => {
+                navigator.geolocation.getCurrentPosition(resolve, reject);
+            });
+
+            // Do something with the latitude and longitude
+        } catch (error) {
+            switch (error.code) {
+                case error.PERMISSION_DENIED:
+                    console.debug("User denied the request for geolocation.");
+                    break;
+                case error.POSITION_UNAVAILABLE:
+                    console.debug("Location information is unavailable.");
+                    break;
+                case error.TIMEOUT:
+                    console.debug("The request to get user location timed out.");
+                    break;
+                default:
+                    console.debug("An unknown error occurred.");
+                    break;
+            }
+        }
+        return {};
+    }
+    checkFilledAll() {
+        this.to_sign_update =
+            _.filter(this.info.items, (item) => {
+                return (
+                    item.required &&
+                    item.role_id === this.info.role_id &&
+                    !SignRegistry.get(item.field_type).check(item)
+                );
+            }).length === 0;
+        this.checkToSign();
     }
 }
 SignOcaPdfPortal.template = "sign_oca.SignOcaPdfPortal";
