+++ sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.xml
@@ -41,4 +41,17 @@
             </div>
         </div>
     </div>
+
+    <t t-name="oca_sign_oca.SignatureButtons" owl="1">
+        <div class="o_sign_oca_buttons" role="toolbar" aria-label="Main actions">
+            <button
+                type="button"
+                class="btn btn-primary o_sign_oca_button_sign"
+                accesskey="s"
+                t-if="to_sign"
+            >
+                Sign
+            </button>
+        </div>
+    </t>
 </templates>
